% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIA_functions.R
\name{calc_S}
\alias{calc_S}
\title{Calculate S generally}
\usage{
calc_S(best_pars, S_fn = calc_S_base, fn_list = NULL)
}
\arguments{
\item{best_pars}{named vector of parameters.  "tau_1" must be a name.  As must "tau_2" and "gamma_A" and "gamma_B"}

\item{S_fn}{function to calculate}

\item{fn_list}{NULL}
}
\value{
Hewlett's S for the given model
}
\description{
Calculate S generally
}
\examples{
best_pars <- c("tau_1" = 0,
              "tau_2" = 1,
              "gamma_A" = 1,
               "gamma_B" = 1)
calc_S_base(best_pars) # should be 1
}
