% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIA_functions.R
\name{calc_S_base}
\alias{calc_S_base}
\title{Calculate S from given tau_1 for base model}
\usage{
calc_S_base(best_pars, fn_list = NULL)
}
\arguments{
\item{best_pars}{named vector of parameters.  "tau_1" must be a name.  As must "tau_2" and "gamma_A" and "gamma_B"}

\item{fn_list}{NULL}
}
\value{
Hewlett's S for the base model.
}
\description{
Calculate S from given tau_1 for base model
}
\examples{
best_pars <- c("tau_1" = 0,
              "tau_2" = 1,
              "gamma_A" = 1,
               "gamma_B" = 1)
calc_S_base(best_pars) # should be 1
}
