% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{design_grid}
\alias{design_grid}
\title{Function to design an experimental grid of combinations}
\usage{
design_grid(
  levels_A = c(0, 1 * 2^(-4:2)),
  levels_B = c(0, 2 * 2^(-4:2)),
  n_rep = 1
)
}
\arguments{
\item{levels_A}{levels of A used in the combination}

\item{levels_B}{levels of B used in the combination}

\item{n_rep}{number of total repetitions of experiment}
}
\value{
data frame with columns dose_A, dose_B, and GIA for all possible combinations
}
\description{
Function to design an experimental grid of combinations
}
