% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_ed_line}
\alias{get_ed_line}
\title{Helper function to get the ED50 line}
\usage{
get_ed_line(
  grid_width = 50,
  par,
  GIA_fn = base_GIA,
  fn_list = NULL,
  ed_val = 50
)
}
\arguments{
\item{grid_width}{number of levels to find points at}

\item{par}{named vector of parameters}

\item{GIA_fn}{function to calculate GIA}

\item{fn_list}{additional parameters to pass to GIA_fn}

\item{ed_val}{Which line to compute.  Default is 50}
}
\value{
data frame with the following columns
\describe{
\item{dose_A}{dose of A (unscaled)}
\item{dose_B}{dose of B (unscaled)}
\item{GIA}{value of GIA \%}
}
}
\description{
Helper function to get the ED50 line
}
