% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rh5_rh4}
\alias{rh5_rh4}
\title{RH5 and RH4}
\format{a 48 x 3 data set where the columns are of the following format
\describe{
\item{RH4}{dose of RH4 in mg/mL}
\item{RH5}{dose of RH5 in mg/mL}
\item{GIA}{Percent Growth inhibition assay averaged over two experiments}
}}
\usage{
rh5_rh4
}
\description{
The data is the raw data for a combination dose of RH5 and RH4.  The data was originally published in Williams et al. (2018).
}
\examples{
data("rh5_rh4")
head(rh5_rh4)
}
\keyword{datasets}
