% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIA_functions.R
\name{SSE_GIA}
\alias{SSE_GIA}
\title{Calculate the Sum of Squared Error}
\usage{
SSE_GIA(par, data, GIA_fn = base_GIA, fn_list = NULL)
}
\arguments{
\item{par}{named vector of parameters}

\item{data}{\itemize{
\item{dose_A}{dose A mg/mL}
\item{dose_B}{dose B mg/mL}
\item{GIA}{GIA}
}}

\item{GIA_fn}{function to calculate GIA}

\item{fn_list}{additional arguments to pass GIA_fn}
}
\value{
sum of square error between observed and estimated
}
\description{
Calculate the Sum of Squared Error
}
