% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cyrpa_ripr}
\alias{cyrpa_ripr}
\title{CyRPA and RIPR}
\format{An object of class \code{data.frame} with 38 rows and 15 columns.}
\usage{
cyrpa_ripr
}
\description{
The data is the raw data for a combination dose of CyRPA and RIPR.
\describe{
\item{well}{ one of iRBC (the max), uRBC (the min), RPMI (??), or comb (which is short for combination)}
\item{RIPR}{dose of RIPR in mg/mL}
\item{CyRPA}{dose of CyRPA in mg/mL}
\item{exp{x}{y}rep{z}}{the results from experiment x, sub item y, repetition z}
}
}
\examples{
data("cyrpa_ripr")
head(cyrpa_ripr)
}
\keyword{datasets}
